Fix thread-safety, correctness, and robustness issues (audit round 2)
CRITICAL:
- VSTChain: add chainLock_ to all reader methods (getPluginSlot,
  getPluginCount, setPluginBypassed, parameter accessors, editor methods)
- VSTChain: make prepared_ std::atomic<bool> (RT/message thread race)

HIGH:
- OutputRouter: clamp numSamples to scaledBuffer_ capacity (overflow guard)
- AudioEngine: use uint32_t for notification queue indices (int overflow)
- WebSocket: case-insensitive HTTP header matching per RFC 7230
- HTTP API: return proper status codes (404, 405, 400) instead of always 200

MEDIUM:
- MidiHandler: dispatch actions outside bindingsMutex_ (deadlock prevention)
- PluginChainEditor: SafePointer in addPluginFromDescription callAsync
- LogPanel: uint32_t ring buffer indices (int overflow)
- VSTChain: resize editorWindows_ before movePlugin to prevent mismatch
- VSTChain: fix processBlock redundant buffer size check

LOW:
- PluginScanner: SafePointer in all 3 callAsync lambdas from scan thread
- Main.cpp: clamp activeSlot to 0-4 in tray tooltip

+++ host/Source/Audio/AudioEngine.cpp
@@ -552,19 +552,19 @@ void AudioEngine::audioDeviceError(const juce::String& errorMessage)
 
 void AudioEngine::pushNotification(const juce::String& msg, NotificationLevel level)
 {
-    int w = notifWriteIdx_.load(std::memory_order_relaxed);
-    int r = notifReadIdx_.load(std::memory_order_acquire);
-    if (w - r >= kNotifQueueSize) return;  // Queue full, drop (RT-safe)
-    notifQueue_[w % kNotifQueueSize] = {msg, level};
+    uint32_t w = notifWriteIdx_.load(std::memory_order_relaxed);
+    uint32_t r = notifReadIdx_.load(std::memory_order_acquire);
+    if (w - r >= static_cast<uint32_t>(kNotifQueueSize)) return;  // Queue full, drop (RT-safe)
+    notifQueue_[w % static_cast<uint32_t>(kNotifQueueSize)] = {msg, level};
     notifWriteIdx_.store(w + 1, std::memory_order_release);
 }
 
 bool AudioEngine::popNotification(PendingNotification& out)
 {
-    int r = notifReadIdx_.load(std::memory_order_relaxed);
-    int w = notifWriteIdx_.load(std::memory_order_acquire);
+    uint32_t r = notifReadIdx_.load(std::memory_order_relaxed);
+    uint32_t w = notifWriteIdx_.load(std::memory_order_acquire);
     if (r == w) return false;
-    out = notifQueue_[r % kNotifQueueSize];
+    out = notifQueue_[r % static_cast<uint32_t>(kNotifQueueSize)];
     notifReadIdx_.store(r + 1, std::memory_order_relaxed);
     return true;
 }
